# Improve performance of JSONPointer encoding