@d-ronnqvist 2024-11-21
> Also, should we entirely remove the `EscapedCharacters` enum? Leaving that in the code might lead to a maintenance issue: If we ever decide to escape other characters someday we might simply add them to the enumeration, not realizing the optimized code here doesn't use that any more. If we leave `EscapedCharacters` in the code, we should at least add a comment explaining that it's not used.

It's public API so we'd need to deprecate it and phase it out. There was never any public API to escape / unescape a `JSONPointer` so there wasn't really a reason for `EscapedCharacters` to be public API to begin with, other than as information about what characters need escaping and what their replacements are. I don't know how useful it is for a client of DocC to have access to that information I didn't feel like taking it away in this PR.